Seperate fine tuning and adapter tuning modules
