+++ plausibility_vaccine/run.py
@@ -1,22 +1,19 @@
 import copy
 import logging
-from typing import Dict, List, Optional, Tuple
+from typing import Dict
 
 import evaluate
 import numpy as np
-from adapters import AdapterTrainer, AutoAdapterModel
+from adapters import AdapterTrainer
 from transformers import (
-    AutoConfig,
-    AutoTokenizer,
     DataCollatorWithPadding,
     EvalPrediction,
-    PreTrainedModel,
-    PreTrainedTokenizer,
     TrainingArguments,
 )
 
+from plausibility_vaccine.adapter import save_delete_adapter, setup_adapters
 from plausibility_vaccine.data import get_data, preprocess_function
-from plausibility_vaccine.fine_tune import save_delete_adapter, setup_adapters
+from plausibility_vaccine.fine_tune import load_pretrained_model
 from plausibility_vaccine.util.args import (
     FinetuningArgument,
     FinetuningArguments,
@@ -37,9 +34,11 @@ def run(
 
     for task_name, task_args in finetuning_args.pretraining_tasks.items():
         logging.info('Running %s', task_name)
-        _run_task(model_args, training_args, task_args)
+        # _run_task(model_args, training_args, task_args)
 
     for task_name, task_args in finetuning_args.downstream_tasks.items():
+        if 'base' not in task_name:
+            continue
         logging.info('Running %s', task_name)
         _run_task(model_args, training_args, task_args)
 
@@ -53,7 +52,7 @@ def _run_task(
     data_args, adapter_args = task_args.data_args, task_args.adapter_args
     raw_datasets, label_list = get_data(data_args)
 
-    model, tokenizer = _load_pretrained_model(model_args, label_list=label_list)
+    model, tokenizer = load_pretrained_model(model_args, label_list=label_list)
 
     with training_args.main_process_first(desc='dataset map pre-processing'):
         raw_datasets = raw_datasets.map(
@@ -108,39 +107,6 @@ def _run_task(
 
     # Save and deactivate the trained adapters to restore the base model
     save_delete_adapter(model, data_args.task_name)
-
-
-def _load_pretrained_model(
-    model_args: ModelArguments, label_list: Optional[List[str]]
-) -> Tuple[PreTrainedModel, PreTrainedTokenizer]:
-    logging.info(f'Loading pre-trained model: {model_args}, label_list: {label_list}')
-
-    config = AutoConfig.from_pretrained(
-        model_args.pretrained_model_name,
-        num_labels=len(label_list) if label_list is not None else 1,
-        cache_dir=model_args.cache_dir,
-    )
-    tokenizer = AutoTokenizer.from_pretrained(
-        model_args.pretrained_model_name,
-        cache_dir=model_args.cache_dir,
-    )
-    model = AutoAdapterModel.from_pretrained(
-        model_args.pretrained_model_name,
-        config=config,
-        cache_dir=model_args.cache_dir,
-    )
-
-    # Some models have set the order of the labels to use, so let's make sure we do use it.
-    if label_list is not None:
-        label_to_id = {v: i for i, v in enumerate(label_list)}
-        model.config.label2id = label_to_id
-        model.config.id2label = {
-            id: label for label, id in model.config.label2id.items()
-        }
-
-    logging.debug('Loaded pre-trained model: %s', model)
-    logging.debug('Loaded pre-trained tokenizer: %s', tokenizer)
-    return model, tokenizer
 
 
 def _run_trainer(trainer: AdapterTrainer) -> None:
